optimise nav rerender
