+++ src/layouts/Layout.tsx
@@ -1,5 +1,5 @@
 import { Outlet } from 'react-router';
-import { NavBar } from '../components/NavBar';
+import { NavBar } from '../components/NavBar/NavBar';
 
 export default function Layout() {
 	return (
